Renderer bug fix.
Fixed bug where the screen flickered between two window sizes. Also
added two types for angles, `degree` and `radian`, so that we get
better type-safety (and because I used degrees in a function that
expected radians, which took a while to figure out). Lastly, matrix
and color extensions to OpenGL has been added.