Example of communication between modules.
The communication is made between the "human interface device" (hid)
module and the graphics module. This is done by pushing a message onto
a queue and reading it at an appropriate time. Instead of having an
event module where everything is sent through, I thought it was best
to let every module implement its own queuing system.