Attempt at handling render components.
Adding new components for rendering is a constant-time operation.
Removing them, or changing their data, however is not as simple and is
currently linear in time.

Also, made a thread-safe queue based on atomics rather than mutexes
and what not. This needs to be complemented with something that blocks
if there is a situation with (for example) multiple writers.

Work in progress. Comments appreciated.