Imported ye old font loading library.
When adding the font I encountered an issue with X11, surprise! Lucky
me. It turns out that when using the newer (not from the 70's)
glx-functions so that we one day may start using OpenGL 4.X, there is
no context available in which we can create the fonts.

Temporary solution is to not use any new OpenGL features (as if we
would anyway, duh). Proper solution is to not rely on old
functionality (a.k.a. glXUseXFont etc.) and instead use shaders
together with freetype or something.

Btw, since I classify fonts as a binary resource, I did not include it
in the commit. Please add it yourself locally until we figure out how
to handle such resources.