Fundamental mathematics
Added types for basic linear algebra as well as quaternions. The types
have been written with the latest advancements of modern c++ in mind
(citation needed). The `get` and `set` methods of the types typically
accept an array where the data should be written to/read from. This is
by design to more easily add support for SIMD instructions in the
future.