More physics attempt
Most of the code is more or less directly taken from Chris Hecker's
old 2D physics sample. There is a lot of things missing; not only
things like collision of two objects that are both moving, but also a
lot of corner cases like edge collisions. Nonetheless, it kind of
works. Physics is definitely possible to do!

A lot of things in the maths library had to change, or rather added
to, in order to get the physics work. Specifically, I had forgotten to
add support for 2D rotation matrices and the like.