Animation is now a thing.
Setting up a model  right in Blender and it is  now possible to export
armatures as a set of joints. Also, meshes can be skinned to armatures
and animated together with them. That is pretty sweet.

Another new thing is the addition  of Chester. Resources should now be
listed in  `chester.txt` and stashed  before committing any  code that
needs them.  This is to  ensure that the program  can be run  from any
commit, but enough about that.