Huge rendering overhaul.
This commit introduces types for entities and assets. Although, so far
there are no assets. The big thing in this commit, however, is the new
collection type that  handles the storage of  components. Currently it
is only used  in the renderer but  should/might be a good  idea to use
elsewhere (that is, for physics and whatnot).