Model decoupling from graphics.
The  armature system  is  now  independent of  the  renderer, yet  the
functionality looks the  same (or possible better!). A  lot of changes
has been  made, particularly  to the collection  class and  its circle
queue counterpart. An improved player model has been added, so be sure
to run chester!

The next thing to do would be to decouple physics also.