Some sort of camera.
The viewer subsystem in the graphics cluster contains all the logic of
cameras, that is of projections and  positioning. The idea was to make
it easy to create and manage cameras via the use of entities. Although
it seems to  work, it turned out  to be a lot more  finicky than first
imagined. This was partly due to missing 4D implementations of vectors
and  the conversions  to  their lesser  dimensional counterparts.  The
vector mathematics needs to be revised.