Kinematic objects is now a thing.
When creating objects; Kinematic is now an option which creates an
dynamic object which is not affected by physics.
Movement update now has a friend Translation update which updates the
position and rotation of an Kinematic object.
Changing sign of rotation when updating graphics which was wrong.
More hardcoded objects, moving platform and rotating beam of doom.