Trigger objects added to physics.
When creating the physics type can be set to trigger. Which will create
an "invisible" collision object that notifies when something enters.
It is also possible to create an oridinary physics object and give it
the behaviour to act as an trigger. It will when provide callback when
something collides in it.