Fixed rotation peculiarities.
The  first issue  was that  quaternions  read from  Blender and  PhysX
needed to be conjugated. This was due to a bug in making matrices from
quaternions. I swear, this was the 10th time I fixed that function.

The second problem was that the  bounding boxes in PhysX got messed up
when playing animations.  There were  many things involved in this bug
but it can be summarized as  PhysX also needing rotations of the boxes
when creating them, not only positions.