Level movement bounds for :camera:
Physics engine now manages movement for Cameras with physical realism.
Three methods are added to physic::camera namespace; one to set the
bounds, one to add a camera and one to update a camera.
Remaining is to load the bounds from the level file and to make the
updating methods threadsafe.