View gravity in layouts.
All views has an gravity attribute which decides how the view will be
positioned in its layout group.
Linear group supports gravity in the direction opposite to its
orientation. So the views in a "vertical" group can have "horizontal"
gravity to make it align to left, center or right.
Relative layout supports anything.