Add support for split screen :cross:
Add a display type to the renderer for showing any number of views of
the scene. As a side effect, the bar rendering had to be disabled
since the old screen transformations are not correct anymore.

Viewer has been updated with frames, splits, and projections. Together
they make it almost easy to create displays. The functions wrapping
the above mentioned transformations have been removed.

Gamestate now defines a setup with three frames as an example to work
with.