Change command into a class :classical_building:
Commands used to be based on an enum which was problematic in the long
run; they have to be defined in one file, meaning we would have to
define gameplay commands in the engine.

Now commands are instances of a class type, based on assets and
hashing of strings. This solves the above problem but means that
commands no longer are nicely enumerated.