Add replay reader :movie_camera:
Added a string structurer that reverses what the string serializer
does, that is it reads data into types from a text string.

Updated the commands serialize function to also support structuring.

Added a replay reader that produces commands and sends them to
gamestate, as dictated by the recording it reads, in this case
"replay.rec".