Make array_alloc trivially destructible :desert:
If the value type of the array_alloc is trivially destructible, it is
never necessary to destruct the elements of it. The array_alloc will
mimic the copy/movability of the value type up to a certain
degree. Not trivially copy/movable value types with a static capacity
can neither be copied nor moved, dynamic capacities cannot be copied,
and although not trivially destructible value types with a dynamic
capacity can be move assigned to care must be taken to appropriately
destruct all elements before the assignment takes place.