Add tuple for generic programming :nut_and_bolt:
So it turns out that the standard tuple is broken, when given
something that cannot be copied and then asking if the tuple can be
copied it will answer yes, because the copy constructor always
exist. Calling said constructor will, however, cause an error (maybe,
I actually never bothered with testing that). There for, this type is
necessary as it will correctly not declare a copy constructor when its
types cannot all be copy constructed.