Change to buffered stream instead :file_cabinet:
The buffered stream is built on top of a read stream, generic enough
to also work for network data. A structurer can wrap a read stream in
a buffered stream if needed for easier text parsing.

In order to support reading of large files, we need the ability to
read more data on request. Originally I had in mind some kind of
coroutine-like thing, but with exceptions (or setjmp), but it was
weird, and after learning what coroutines are it became clear to me
that that was not exactly what I wanted either. In the end, a simple
callback for reading more data turned out to be pretty clean. Maybe a
class with a single virtual function might be even more so?