Use read stream with json :coffee:
The nlohmann library is a bit awkward to read data into, the only way
to not give it all data at once appears to be to use
`std::stream`. That is not generic enough for this new unified
streaming interface we have so we default to the bad way of giving it
everything.