Replace array_alloc with array_wrapper :repeat:
The array_alloc turned out to be somewhat difficult to work with, so
instead a better replacement was made. The only real difference
between the two is that the wrapper also contains the size and not
only the capacity. This makes the wrapper more complete, actually
helps remove some duplicated code, and otherwise more easy to work
with.