Add shader to freetype fonts :a:
Fonts (with freetype) have been updated to contain distance
fields. This makes it possible to do cool shader tricks to get super
sharp text, and other things like shadowing.

Rewrote the reading of fonts (with freetype) to be on the same format
as our other file readings. This helps streamline our code and test
the limits of our serialization framework.

In order to keep things simple all freetype code was moved into the
renderer, obviously not an ideal solution. In fact, the technique of
distance fields can and should be applied to the other types of font
loading we support, thus the freetype code will by necessity have to
move out again.