Make input platform independent :innocent:
Rewrote the input type to make it easier to add more different kinds
of input to it, like gamepads.

The input event translations are broken right now, although windows is
slightly less broken than linux :/ Work has been made to make this
work on linux but it is not included in this commit due to the volume
of it and it mostly being x11 stuff. Also, it would be nice to make it
work on windows first in case things have to adjust.