Massage msvc to do our bidding :massage:
For whatever reason, the function specialized in `contiguous_iterator`
and the function with `decltype` was according to msvc equally
specialized so we got an ambiguity error here. By changing the ints to
signed and unsigned we, together with the ellipses, effecitvely make a
three-way comparison. This helps disambiguate the problem for the
compiler. Maybe it helps a bit with readability as well!