Add unicode data to key character inputs :unicorn:
X11 has a lot of strange addons to it, and input method is one of
them. Reading about it online, people seem to think it is deprecated
and superseded by something called ibus or something. Anyway, input
method is good enough for now, it does handle the swedish locale
correctly with dead keys and all.