Add mapping functionality to devices :world_map:
A filter is a thing that recognizes certain types of input, the type
of the filter controls the logic that should be executed.

A mapping is a particular set of filters and can be bound to states in
a context. The context then will make sure to make this mapping
available for the devices it is connected to. Conflicts in the device
mapping occurs when multiple overlapping mappings are active at the
same time, irregardless to which context they belong.