Improve on debug breaking :stop_sign:
By moving the debug breaking out from the functions and into a macro,
the debugger will in most cases now actually point to the line of code
with the issue, rather than pointing into an underlying
function. However, any assembly we use have to be put in a function
anyway so when that is used we will not get the benefit of this.