Replace command posting with callback :arrow_right_hook:
This change enables a couple of things. In particular this decouples
the ui from gameplay, solving the problem of how the engine can
communicate with gameplay. This also makes it possible to extend the
input system easily without having to make changes to ui, thus it can
be made smaller and be given a clear one responsibility. The cost is
some extra memory and function pointers. The extra memory can be
mitigated by storing the callbacks separately.