Replace display hack with poll :raised_hands:
When looking into writing a test for the window creation and
destruction, the call to `XSelectInput` started failing with a
`BadWindow` message. The cause was most probably because the window
argument was not a window of the given display. A similar problem can
be read about here:

https://stackoverflow.com/q/11272054

This whole multiple display situation was a hack to separate the
window message polling from the graphics drawing. It was not until
this problem that I stumbled upon the `ConnectionNumber` macro, as was
the solution to yet another similar problem:

https://stackoverflow.com/a/8592969

After this rewrite, all X11 code runs from one thread, except for when
the render thread wants to butt in and swap some buffers.