Remove gameplay dependencies from engine :game_die:
If we remove the gamestate parameter from `post_command` function it
becomes a lot easier to use as a callback. It makes sense to not have
to pass the gamestate around since the gamestate module should be able
to figure things out simply from the entity it is being passed.