Not remove binary because tests fail :hugging_face:
Previously when tests failed they would cause the test binary to
fail (even though it must have existed because we ran the tests). This
was annoying and prevented us from debugging failing tests. This
change fixes this by manually keeping track of passing tests so that
the custom target they are attached to do not run them every time we build.