Inline debugger with assembly break :inbox_tray:
Without this fix, the debugger will point to the inline function
containing the assembly rather then the function using the macro. This
is bad user experience for clearly they want to see their own code,
not this function! It turns out that the debugger points to the
correct line when using a lambda instead.