Fix bug with growing allocations :bug:
Before, it could be the case that the amount was less than 8 but the
sum of capacity and amount was greater than 8. This was detected when
concatenating two strings and the result only contained the first
string.

The suggested fix of using the ceiling function makes it no longer
possible to request growths of zero size.