Add inotify file system :interrobang:
Only the watch function depends on inotify, everything else is
posix. It might be good to separate the two so that some other
implementation can reuse the posix stuff.

The tests/game cannot be run with the dummy implementation as the
synchronization primitives that are used to wait for the file module
will never trigger. This should probably be fixed at some point.