Add license :scroll:
So why ISC? The main reason why is simply that I do not know anything
about licenses. A secondary reason is that I want this particular
piece of code to be open source, while still leaving some wiggle room
to allow for other closed source projects to build on this.

GPL 3 was rejected due to its length and scariness. Also this:

https://www.youtube.com/watch?v=PaKIZ7gJlRU

GPL 2 might be interesting, but again it is too much to read for my
little brain.

MIT is basically ISC but one sentence more.

No other licenses where considered. Do you feel strongly about
licenses and would like to change this one? Contact us!