Add native file reading support :nerd_face:
This is a low level interface to interact with files, and particularly
for this commit about file reading. Most file usage should probably go
through the file module in the engine library instead of using this
interface directly.

One notable exception where this interface actually could be the
better choice is at start up. For instance, the main application might
want to read settings from files before starting any engine modules,
and since the file module is an engine module we have something like
the bootstrapping problem. This then would be an easy workaround.