Delay execution of printing arguments :timer:
Since c++ does not guarantee the order in which function arguments are
evaluated (other than that c++17 promises not to interleave them), the
assert macro could lie about the cause of failure. This turned out to
be the case on windows with msvc where `GetLastError` sometimes was
evaluated before the function call that generated the error.

This is not a problem on posix systems that use the `errno` variable
since it will be passed as a reference and always read from after the
erroneous function has returned. However, the underlying issue with
evaluation order remains so this fix is still necessary.