Make completion cleanup after itself :broom:
Allowed `CancelIoEx` to fail nicely. This happens when the function is
called after the kernel has read directory changes and posted an apc
message about it but before we have had a change to read it. As a
result of this delay, the cleanup of async data has been completely
moved to the completion function. Hopefully, it will ALWAYS be called
:worried: