Remake how materials work :hammer:
Seperate materials into a material class and a material instance.
The classes contain data common for all materials sharing the same
class while the instances contain more detailed data and referes to
its class. In the case of overlapping data between the class and
instance, the class has the precedence.