Add fragmented container :bar_chart:
Similar to arrays, the fragmented data type stores its elements in an
array. Its elements are, however, never moved to "fill in the holes"
when an element is removed, hence its name `fragmentation`.

In order for the container to keep track of what indices are free to
use for new elements an extra index buffer is appended to the
storage. This buffer is completely private and not exposed to users.

Due to its fragmented property it is not possible to iterate over its
elements. Technically it could be changed to support this but then
adding/removing would probably have to be changed from being constant
time algorithms.