Rework container hierarchy :deciduous_tree:
Made some adjustments to the container hierarchy that had surprisingly
big effects on its users. In particular, this change made many
simplifications possible, and also corrected some implementation flaws
about constructing objects correctly.

The newly added `good` check on storages is only valid to call after a
previous call to `allocate` but prior the call to `deallocate`. It
returns the status of the allocation and is more correct to call than
the previous check for static capacity since even storages with static
capacity might have to deallocate.