Add zip pointer type :point_down:
Iterators are not required to be comparable with `nullptr`, but
pointers are. According to some people, however, pointers are not
needed to support arithmetic operators. Quite frankly, the whole
situation is quite messy.

The document D0773R1 for the Library Evolution Working Group sheds
some light to this mess that is "fancy pointers" although it does not
deal with "fancy" proxy pointer types :disappointed: