Limit the dependency configurations :no_good:
Previously dependencies were built with the same flags as the main
project. Thus there was a one to one mapping between the main project
and dependencies.

Now we limit the dependencies to only support a "release" and "debug"
configuration, controlled by the variables `DEP_RELEASE_CONFIG` and
`DEP_DEBUG_CONFIG`. Since we now no longer have a one to one mapping,
the configuration is chosen as follows. If the debug build is
enabled (that is if `DEP_DEBUG_CONFIG` has a value) and your main
configuration is listed in the `DEBUG_CONFIGURATION` property then the
debug build will be used. Otherwise, the release build will be used.