Fix issue with recursive begin/end :recycle:
The begin/end functions in the "utility" namespace will be picked up
as candidates for any type in the same namespace. The solution is to
limit these functions to only work on ranges, as defined to support
standard begin/end.

This problem was found when trying to upgrade the Catch
library. Clearly, they do some cool type inspection stuff.