Add support to build Coz :building_construction:
Coz is put in the tools directory since it strictly is not necessary
to run Fimbulwinter, even though it technically adds a dependency to a
header file.

The optional patch step is necessary for systems that link against
glibc, read the man pages for gettid(2).

Also changed the compile option from `-Og` to `-g` since the former
did not work when running Coz.