Build dependencies individually :woman:
Checking config in the generated build script for dependencies turned
out to not be possible with Visual Studio, the `CMAKE_CFG_INTDIR`
could not be used in the configure context like that.

This change tries to remedy that by wrapping each dependency in its
own build script, thus removing the need to check which config we are
building for, and instead assume that all available configs should be
built. A consequence of this is that the `ExternalProject_Add`
function is slightly less intuitive to use as we now have to use a
wrapper function instead.