Attempt at beating standard vector :punch:
Added two alternatives to `try_emplace_back`, controlled by tag
distpatching on `no_reallocate` and `no_failure`. The former is not
strictly better than the regular version but remains useful as it
shows intent. The latter, however, does show an improvement, and for
small ranks even competes with c-style copying.