Disable conversion to any when printing :no_entry_sign:
Since anything (basically) can be implicitly converted to any, and any
has a print function, it has meant that anything can be printed. This
started to become a problem whith string iterators in asserts in that
is_ostreamable could succeed yet an error would trigger in
try_stream_yes_t.

The fix then is to either disable implicit conversions to any, which
would be a difference to how the standard any works, or disable
conversions just for the printing function. This change is an attempt
at the latter.