Restructure file system and add loader :recycle:
The file loader builds on the file system and takes care of only
reading files that has not been loaded previously. It scans
directories for files and puts them in a table based on the hash of
their names. This makes it possible for users to load files based on
(hashed) numbers instead of relying on file paths.

The usage of libraries was inspired by the virtual file system library
PhysicsFS.