Add dummy task scheduler :calendar:
The need of a task scheduler came from working out how to support file
watching well. Badly it is easy to support but if we want guarantees
over file re/unloading with minimal impact on the other modules with
respect to resource creation/removal then we better enforce when said
re/unloading occurs.