Guard against multiple reads without changes :guardsman:
The FILE_ACTION_MODIFIED action on file watches is being recieved
multiple times for the same file write. Not only does it feel
unnecessary to read the file twice without any changes have been made
in between, but it also is problematic if the user does not expect the
callback to be called more than once for one file change.