[emoji] Separate file type reading and load requests
Each filetype has two callbacks, one for loading (which reads the
file), the other for unloading. Reloading a file because of file
watching will simply call the loading callback. This means unloading
is only ever done when no resources use the file anymore.

Each load request also has two callbacks, one for signaling
readyness (when everything is loaded), the other for unreadyness.

This implementation does not guarantee the order in which the
callbacks are called when reloading/unloading. To fix this we either
need to add more synchronization primitives, e.g. locks and stuff, or
some organized threading system.