Add shared pointers :handshake:
The standard shared pointer is good but unnecessarily large. Our
implementation is only one pointer in size and does not support
elevating arbitrary pointers to be shared.

The reduced object size is welcomed if used together with the any
type, where we can avoid an extra allocation that the standard shared
pointer otherwise causes.

Arguably the standard shared pointer is faulty since it supports
assigning arbitrary pointers. Using make_shared is clearly superior.