Report existing and removed files separately
While in the process of splitting the text string for file changes, it
turned out useful to also split the collection for files. This change
adds a new collection for loaded (or in the process of being loaded)
files while the previous collection remains for keeping track of the
currect state of the filesystem.