Change file load data to a stash
The way the user data worked with file loading (not file
readyness!) was strange, not very well thought out, and buggy. This
change replaces that parameter with a stash the user is free to store
anything in and can be access when unloading as well as from the
readyness callbacks.