Separate constexpr Hash from Asset :hash:
Hash objects can be thought of as something static that only lives
inside the code. Asset objects, however, can be thought of as
something dynamic that usually comes from outside the code, such as
from files. That is why Hash cannot be serialized but Asset
can.

Similarly, Command objects are a strange mix that can do both. Maybe
this should be revised :thinking: